rag_chatbot/requirements.txt
groq>=0.9.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pandas>=2.2.0
pdfplumber>=0.11.0
streamlit>=1.35.0
requests>=2.32.0
scikit-learn>=1.5.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0
matplotlib>=3.9.0
plotly>=5.22.0
tqdm>=4.66.0
